# Conflicts:
#	src/main/java/com/example/data22aexamprojectlnl2/controllers/PosterController.java
#	src/main/java/com/example/data22aexamprojectlnl2/models/Company.java

+++ src/main/java/com/example/data22aexamprojectlnl2/controllers/PosterController.java
@@ -1,7 +1,9 @@
 package com.example.data22aexamprojectlnl2.controllers;
 
+import com.example.data22aexamprojectlnl2.models.Company;
 import com.example.data22aexamprojectlnl2.models.Image;
 import com.example.data22aexamprojectlnl2.models.Poster;
+import com.example.data22aexamprojectlnl2.models.Security;
 import com.example.data22aexamprojectlnl2.repositories.ImageRepository;
 import com.example.data22aexamprojectlnl2.services.ImageService;
 import com.example.data22aexamprojectlnl2.services.PosterService;
@@ -11,6 +13,7 @@ import com.example.data22aexamprojectlnl2.repositories.PosterRepository;
 import com.example.data22aexamprojectlnl2.services.ImageService;
 import com.example.data22aexamprojectlnl2.services.PosterService;
 
+import com.example.data22aexamprojectlnl2.services.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -43,7 +46,9 @@ public class PosterController
     @Autowired
     ImageRepository imageRepository;
 
-    // flyt til image controller
+    @Autowired
+    SecurityService securityService;
+
     @GetMapping("/getImages")
     public ResponseEntity<List<Image>> getAllImagesByPosterId(@RequestParam("poster_id") int poster_id)
     {
@@ -53,17 +58,24 @@ public class PosterController
 
 
     @DeleteMapping("/deletePoster")
-    public ResponseEntity<String> deleterPosterByPosterId(@RequestParam("poster_id") int poster_id) {
-        Optional<Poster> checkPoster = posterService.getPosterById(poster_id);
-        if(checkPoster.isPresent()) {
-            posterService.deletePoster(poster_id);
-            return ResponseEntity.ok("Poster deleted");
+    public ResponseEntity<String> deleterPosterByPosterId(@RequestParam("poster_id") int poster_id,
+                                                          @RequestParam("username") String username,
+                                                          @RequestParam("password") String password) {
+        String hashedUsername = passwordHashing.doHashing(username);
+        String hashedPassword = passwordHashing.doHashing(password);
+        Optional<Security> checkSecurity = securityService.getSecurityByUsernameAndPassword(hashedUsername, hashedPassword);
+        if(checkSecurity.isPresent()) {
+            Optional<Poster> checkPoster = posterService.getPosterById(poster_id);
+            if (checkPoster.isPresent()) {
+                posterService.deletePoster(poster_id);
+                return ResponseEntity.ok("Poster deleted");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Poster not found");
+            }
         } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Poster not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Wrong username or password");
         }
     }
-
-
 
     // dette er faktisk en /postPost endpoint. der skal tilføjes Description og Titel.
     @PostMapping("/createPost")
